+++ core-bundle/src/Event/PreviewUrlConvertEvent.php
@@ -1,0 +1,46 @@
+<?php
+
+/*
+ * This file is part of Contao.
+ *
+ * Copyright (c) 2005-2015 Leo Feyer
+ *
+ * @license LGPL-3.0+
+ */
+
+namespace Contao\CoreBundle\Event;
+
+use Symfony\Component\EventDispatcher\Event;
+
+/**
+ * Allows to convert a preview URL.
+ *
+ * @author Leo Feyer <https://github.com/leofeyer>
+ */
+class PreviewUrlConvertEvent extends Event
+{
+    /**
+     * @var string
+     */
+    private $url;
+
+    /**
+     * Returns the URL.
+     *
+     * @return string The URL
+     */
+    public function getUrl()
+    {
+        return $this->url;
+    }
+
+    /**
+     * Sets the URL.
+     *
+     * @param string $url The URL
+     */
+    public function setUrl($url)
+    {
+        $this->url = $url;
+    }
+}